feat(gc): a separate module for gc tests
