+++ nightly/nightly.txt
@@ -116,12 +116,12 @@ expensive nearcore test_cases_testnet_rpc test::test_delete_access_key_with_allo
 # expensive nearcore test_cases_testnet_rpc test::test_access_key_smart_contract_testnet
 
 # GC tests
-expensive near-chain store tests::test_gc_remove_fork_large
-expensive near-chain store tests::test_gc_not_remove_fork_large
-expensive near-chain store tests::test_gc_boundaries_large
-expensive near-chain store tests::test_gc_random_large
-expensive near-chain store tests::test_gc_pine
-expensive near-chain store tests::test_gc_star_large
+expensive near-chain gc tests::test_gc_remove_fork_large
+expensive near-chain gc tests::test_gc_not_remove_fork_large
+expensive near-chain gc tests::test_gc_boundaries_large
+expensive near-chain gc tests::test_gc_random_large
+expensive near-chain gc tests::test_gc_pine
+expensive near-chain gc tests::test_gc_star_large
 
 # other tests
 expensive nearcore test_simple test::test_2_10_multiple_nodes
